Move `WasmFeatures` to a standalone module in `wasmparser`
This makes the type available even without the `validate` feature since
it doesn't have any intrinsic dependencies on validation.
